backend/requirements.txt
tensorflow==2.20.0
numpy==1.26.4
Pillow==10.2.0
